reduce warnings.

+++ CrystalAlgo/CrystalPhysics/CSPHBoundarySolver.cpp
@@ -34,10 +34,10 @@ Vector3df CSPHBoundarySolver::getForceX(const float x)
 {
 	float over = 0;
 	if (x > boundary.getMaxX()) {
-		over = x - boundary.getMaxX();
+		over = x - static_cast<float>(boundary.getMaxX());
 	}
 	else if (x < boundary.getMinX()) {
-		over = x - boundary.getMinX();
+		over = x - static_cast<float>(boundary.getMinX());
 	}
 
 	const float force = getForce(over);
@@ -48,10 +48,10 @@ Vector3df CSPHBoundarySolver::getForceY(const float y)
 {
 	float over = 0;
 	if (y > boundary.getMaxY()) {
-		over = y - boundary.getMaxY();
+		over = y - static_cast<float>(boundary.getMaxY());
 	}
 	else if (y < boundary.getMinY()) {
-		over = y - boundary.getMinY();
+		over = y - static_cast<float>(boundary.getMinY());
 	}
 	const float force = getForce(over);
 	return Vector3df(0, 1, 0) * force;
@@ -61,10 +61,10 @@ Vector3df CSPHBoundarySolver::getForceZ(const float z)
 {
 	float over = 0;
 	if (z > boundary.getMaxZ()) {
-		over = z - boundary.getMaxZ();
+		over = z - static_cast<float>(boundary.getMaxZ());
 	}
 	else if (z < boundary.getMinZ()) {
-		over = z - boundary.getMinZ();
+		over = z - static_cast<float>(boundary.getMinZ());
 	}
 	const float force = getForce(over);
 	return Vector3df(0, 0, 1) * force;
